Added a way to execute custom RxBleRadioOperation
The RxBleConnection interface has now a new method named queue that enable
library consumer to queue their own RxBleRadioOperation. The queue method
receives in parameter a RxBleRadioOperationCustom implementation. This
said implementation receives the underlying Android BLE API objects and
is responsible of returning an Observable<T> that executes the operation
asynchrounously.

The Observable<T> returned is then executed as part of normal queue
processing logic. The queue is released when the Observable<T> terminates
either via an error or normally via complete.

This feature shall be used only by people that studied the RxAndroidBLE
source code and understand the implication of running their own operation.