Fixed connection mock direct writeCharacteristic not callind setValue
The other variants of write writeCharacteristic are calling setValue to
change the internal value to this. In some of my tests, I was before
using the UUID version and was relying that setValue would be called acting
as a kind of holder for the read.

This assumption was broken with the writeCharacteristic variant that takes
a BluetoothGattCharacteristic and a byte array.

Resolved this minor thing by also calling setValue on the received
BluetoothGattCharacteristic.