Fix for radio not being released on operation complete. (#246)
https://github.com/Polidea/RxAndroidBle/issues/244
There was an issue (race condition) with releasing radio in `.onTerminate()` inside `RxBleRadioOperation.run()`. If the `Observable` called `.onNext()` and `.onCompleted()` in lines next to each other and user would use `observable.first()` then the `.onComplete()` may not be noticed by the `Subscriber` and `RadioReleaseInterface.release()` in `.onTerminate()` would not be called.