Separated command queues between connections (#260)
(#250) It was found that commands from different `BluetoothGatt`’s are synchronized so it is possible to decouple synchronization of commands for different connections. Thanks to that if there are two or more connections opened and one connection would get stalled — other connections may continue to work without stalling — with a single command queue shared between connections it was not possible.