Fix MTU not updated when changed by the peripheral. (closes#293) (#294)
* Fix `RxBleConnection.getMtu()` returning default value when MTU was changed by the peripheral. (closes#293)
Previously the `RxBleConnection.currentMtu` was updated only when the MTU was updated by the central. This prevented Long Writes from utilising the increased Maximum Transfer Unit or the used from having a valid current MTU value—this made the utilisation of available bandwidth suboptimal.
Introduced `ConnectionSubscriptionWatcher` interface for `MtuWatcher` and `DisconnectAction`. In the future `ConnectionOperationQueue` will also use it probably.