Call ConnectionSubscriptionWatcher on Bluetooth Callbacks Scheduler (fixes #308) (#315)
Call ConnectionSubscriptionWatcher on Bluetooth Callbacks Scheduler (fixes #308)
Without specifcation of `.subscribeOn()` and `.unsubscribeOn()` there was a possibility of a race condition when calling `.doOnSubscribe()` and `.doOnUnsubscribe()` which lead to calling `ConnectionOperationQueueImpl.onConnectionUnsubscribed()` before the `.onConnectionSubscribed()` has returned leading to a `NullPointerException`. By default `.doOnSubscribe()` is called on a thread that calls `.subscribe()` and `.doOnUnsubcribe()` is called on thread that calls `Subscription.unsubscribe()`.