Fixed stalled client queue when adapter is off (#522)
There was a race condition in ConnectorImpl which was simultaneously scheduling connecting operation and observing BT adapter being switched off. As the client operation queue runs on it’s own thread it was possible that the connecting operation would get scheduled and started but cancelled before the actual `protectedRun()` would get executed — this result in a situation where the client queue would get blocked and never released due to the operation never being run but still responsible for releasing.
Workaround: start observing BT adapter state after the connecting operation finishes. The operation is observing it either way during the execution and will emit an error if BT is off