Added option to disable Location Providers check before scan (#533)
Some Android devices are not returning any results if Location Providers (Location Services) are turned off. The library defaults to check if Location Services are turned on before the scan and emitting an exception if not so the user would get a clear information what they may expect. However now the user may start handling this on their own by switching off the check.