Fixed RxBleClient.observeStateChanges() emitting while turning BT on (#556)
Android BluetoothAdapter has 4 states. ON, OFF, TURNING_ON, TURNING_OFF. Any state other than ON evaluates to BLUETOOTH_NOT_ENABLED. For instance when the state observing starts in OFF state and adapter is turned on then a TURNING_ON state is emitted but it is not a change from the subscriber’s perspective.