Added Connection Parameters Update Callback (#565)
The access to the callback is available through `RxBleConnection.observeConnectionParametersUpdates()`

Additionally it is possible to get it in `RxBleCustomOperation`:
- `RxBleGattCallback.getConnectionParametersUpdates()`
- `RxBleGattCallback.setHiddenGattCallback(HiddenBluetoothGattCallback)`

The callback is not used for completing `RxBleConnection.requestConnectionPriority()`. It is not possible to universally distinguish which callback is the one that should complete the request. Why? Android OS does change connection parameters on its own at the beginning of the connection to speedup service discovery process. After the discovery parameters are relaxed. If a request is made during this period it is not possible to know which callback is related to the user's request. It is worth noting that making a request early will make it quickly overwritten by the above mentioned process.

This is _not_ a part of public Android API and is not guaranteed to be called.