Workaround scan failure for non-default callback type and empty filters
Android API may be requested to match first or the last appearance of peripherals. When `ScanFilter`s are not specified or they are empty (match all advertisements) — the hardware/offloaded filtering will likely overwhelm the BT chip — this is why scan is not started with `SCAN_FAILED_FEATURE_UNSUPPORTED` or `SCAN_FAILED_OUT_OF_HARDWARE_RESOURCES` error in the result.

This commit adds a workaround by falling back to emulated callback type in the described scenario — similarily as it is working on API 21 and API 18.