Fixes stalled operation queue in a rare race situation
`.subscribeOn()` operator seems to be not actually subscribing to the upstream if downstream is already disposed. It seems impossible to track whether the upstream was subscribed or not and therefore side-effects for releasing a queue cannot be reliably run. Getting rid of the operator but running the subscription on the `Scheduler`’s worker directly makes the operation being always run. It’s the operation’s responsibility to release the queue at a proper moment (when the `BluetoothGatt` is no longer blocked).