Added check BluetoothAdapter.isEnabled() before calling BluetoothLeScanner.stopScan(ScanCallback)
Apparently on some versions of Android OS it is possible to get BluetoothLeScanner even if bluetooth is off. In this situation calling any methods/functions on the scanner will fail with an IllegalStateException due to adapter being off. Previously the scanner was not possible to obtain when the adapter was off or the exception was not thrown.