Added scan permission helper functions (#642)
Since API 23 (Android 6 / Marshmallow) runtime ACCESS_COARSE_LOCATION or ACCESS_FINE_LOCATION permission is needed for an application to run a BLE scan. Since API 29 (Android 10) only ACCESS_FINE_LOCATION is accepted to run a BLE scan. Added helper functions for checking if any more runtime permissions are needed to run a scan and to get a list of acceptable/recommended runtime permission names — only one of the list is needed to run a scan.