Fixed requirements for location services check. (#748)
Apparently on >= API 29 Android will require location to be turned on irregardless of the target SDK for the application to return scan results. Prior to API 29 it was a common workaround to use target SDK of 21 so the user will not be prompted to give location permissions and no need for location being turned on by the user to scan for BLE peripherals.