Fixed Gradle deprecation warning in CheckerLocationProvider.
Android Studio does honor a comment `//noinspection deprecation` but Gradle continues to complain. To silence Gradle we need to use `@SuppressWarnings`. We want to suppress on the smallest possible scope to not be blind about next deprecations. After refactoring to a separate function and adding `@RequiresApi()` Android Studio does not report deprecations but Gradle still complains. To fix Gradle we need to add `@SuppressWarnings("deprecation")` but the Android Studio complains about redundant deprecation. To silence that we need to add "RedundantSuppression". Looks like a bug.