Fixed CharacteristicLongWriteOperation defaults (#847)
On Android 13 when negotiated MTU > 515, CharacteristicLongWriteOperation would by default split data into chunks that did not fit Android internal data buffer for BLE writes. Reason for that is that on Android 13 the maximum attribute data buffer length was reduced from 600 to 512 bytes - which is now according to BLE Specs. Android 14 makes MTU negotiation to the maximum supported value on Android (517) by default, making this problem more likely to surface.